add activity routes and endpoints for daily, weekly, monthly, and yearly activity tracking
