+++ src/index.ts
@@ -13,7 +13,13 @@ app.use(express.json());
 mongoose.connect(process.env.MONGODB_URI);
 
 const userRoutes = require("./routes/user");
-app.use(userRoutes);
+app.use("/user", userRoutes);
+
+const activityRoutes = require("./routes/activity");
+app.use("/activity", activityRoutes);
+
+const activitiesRoute = require("./routes/activities");
+app.use("/activities", activitiesRoute);
 
 app.get("/", (req, res) => {
   res.status(201).json("welcom on the app");
